feat(side-bar): addresses comments for the side bar
