+++ packages/app/src/main.tsx
@@ -3,9 +3,12 @@ import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 
 import { App } from './App';
+import { Sidebar } from './systems/Core/components/Sidebar';
 
 createRoot(document.getElementById('root')!).render(
   <BrowserRouter basename={import.meta.env.BASE_URL}>
-    <App />
+    <Sidebar>
+      <App />
+    </Sidebar>
   </BrowserRouter>
 );
