+++ packages/app/src/systems/Core/components/Layout/Layout.tsx
@@ -8,9 +8,6 @@ import { Helmet } from 'react-helmet';
 import { BottomBar } from './BottomBar';
 import { TopBar } from './TopBar';
 
-import { Sidebar } from '~/systems/Sidebar';
-import { useSideBar } from '~/systems/Sidebar/hooks/useSidebar';
-
 type Context = {
   isLoading?: boolean;
   isHome?: boolean;
@@ -55,8 +52,6 @@ export const Layout: LayoutComponent = ({
   title,
   children,
 }: LayoutProps) => {
-  const { isSideBarOpen, toggle } = useSideBar();
-
   const titleText = title ? `${title} | Fuel` : 'Fuel';
   return (
     <ctx.Provider value={{ isLoading, isHome, title }}>
@@ -67,19 +62,7 @@ export const Layout: LayoutComponent = ({
         {isPublic ? (
           <>{children}</>
         ) : (
-          <Flex css={styles.wrapper}>
-            {children}
-            <Sidebar
-              size={220}
-              side={'right'}
-              type="tree"
-              isDismissable
-              onClose={() => toggle()}
-              shouldCloseOnBlur
-              shouldCloseOnClickAway
-              isOpen={isSideBarOpen}
-            />
-          </Flex>
+          <Flex css={styles.wrapper}>{children}</Flex>
         )}
       </Flex>
       {import.meta.env.NODE_ENV === 'test' && (
