+++ packages/app/src/systems/Core/components/Layout/TopBar.tsx
@@ -6,6 +6,7 @@ import {
   IconButton,
   Spinner,
   Text,
+  Drawer,
 } from '@fuel-ui/react';
 import type { ReactNode } from 'react';
 import { useNavigate } from 'react-router-dom';
@@ -19,7 +20,6 @@ import { useLayoutContext } from './Layout';
 import { NetworkDropdown } from '~/systems/Network/components/NetworkDropdown';
 import { useNetworks } from '~/systems/Network/hooks';
 import { NetworkScreen } from '~/systems/Network/machines';
-import { useSideBar } from '~/systems/Sidebar/hooks/useSidebar';
 
 type TopBarProps = {
   onBack?: () => void;
@@ -27,7 +27,6 @@ type TopBarProps = {
 };
 
 export function TopBar({ onBack }: TopBarProps) {
-  const { toggle } = useSideBar();
   const { isLoading, title, isHome } = useLayoutContext();
   const { networks, selectedNetwork, handlers } = useNetworks({
     type: NetworkScreen.list,
@@ -71,13 +70,14 @@ export function TopBar({ onBack }: TopBarProps) {
         variant="link"
         css={{ px: '0 !important' }}
       />
-      <IconButton
-        onClick={() => toggle()}
-        icon={<Icon icon="List" color="gray8" size={24} />}
-        aria-label="Open menu"
-        variant="link"
-        css={{ px: '0 !important' }}
-      />
+      <Drawer.Trigger>
+        <IconButton
+          icon={<Icon icon="List" color="gray8" size={24} />}
+          aria-label="Menu"
+          variant="link"
+          css={{ px: '0 !important' }}
+        />
+      </Drawer.Trigger>
     </Flex>
   );
 }
