+++ packages/app/src/systems/Core/components/Sidebar/components/NetworkSelector/NetworkSelector.test.tsx
