+++ packages/app/src/systems/Core/components/Sidebar/constants/index.ts
@@ -1,0 +1,1 @@
+export * from './sidebarItems';
