+++ packages/app/src/systems/Core/components/Sidebar/index.tsx
@@ -1,3 +1,2 @@
 export * from './components';
 export * from './Sidebar';
-export * from './machines';
