--- packages/app/src/systems/Sidebar/components/constants/index.ts
@@ -1,1 +1,0 @@
-export * from './routes';
