--- packages/app/src/systems/Sidebar/machines/index.tsx
@@ -1,1 +1,0 @@
-export * from './sidebarMachine';
